New !word goal command
Computes the progress you should be at in order to meet your goal for the month.

While super handy for NaNoWriMo participants, it makes no assumptions about what month you are writing in, and instead always uses the current month, including calculating daily progress goals based on the correct number of days for the current month.